% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK_mles.R
\docType{data}
\name{AK_mles}
\alias{AK_mles}
\title{MLE's obtained by Aaron King after fitting the He10 model.}
\format{
\subsection{\code{AK_mles}}{

A 20 x 19 tibble. All columns after the first 5 are estimated parameters.

\describe{
\item{town}{City name}
\item{loglik}{Estimate of log likelihood evaluated at the given parameters}
\item{loglik.sd}{Estimated standard deviation for the
log likelihood evaluated at the given parameters}
\item{mu}{Rate at which people die}
\item{delay}{Delay in years before born individuals transition to S}
\item{sigma}{Rate at which people transition from E to I}
\item{gamma}{Rate at which people transition from I to R}
\item{rho}{Probability that person transitioning from I to R is recorded}
\item{R0}{Expected number of people infected by an individual in a
population where everyone is susceptible}
\item{amplitude}{Controls difference in rate at which people
transition from S to E during school versus during vacations}
\item{alpha}{A mixing parameter}
\item{iota}{Mean number of infected individuals visiting the town at any
time}
\item{cohort}{Controls proportion of the year's susceptible individuals
who are added on the 251st day of the year}
\item{psi}{Measurement overdispersion parameter}
\item{S_0}{Proportion of individuals at time 0 who are susceptible}
\item{E_0}{Proportion of individuals at time 0 who are exposed}
\item{I_0}{Proportion of individuals at time 0 who are infectious}
\item{R_0}{Proportion of individuals at time 0 who are recovered}
\item{sigmaSE}{Standard error of gamma white noise applied to rate of
people transitioning from S to E}
}
}
}
\source{
\url{https://kingaa.github.io/sbied/measles/index.html}
}
\usage{
AK_mles
}
\description{
MLE's obtained by Aaron King after fitting the He10 model.
}
\keyword{datasets}
