% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK_model.R
\name{AK_model}
\alias{AK_model}
\title{Create panelPomp object using Aaron King's re-fitting of the He10 model.}
\source{
\url{https://kingaa.github.io/sbied/measles/index.html}
}
\usage{
AK_model()
}
\value{
panelPomp object.
}
\description{
Aaron King's model is like He et al. 2010 model EXCEPT that
population and births are interpolated using smooth splines and the
interpolated covariates are shifted +0.5 years. In contrast, He et al. use
linear interpolation and make no shift. In addition, Aaron King's MLE's are
different.
}
\examples{
\dontrun{
AK_model()
}
}
