% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK_pparams.R
\docType{data}
\name{AK_pparams}
\alias{AK_pparams}
\title{Parameters for \code{AK_model()} in \code{pparams()} form}
\format{
\subsection{\code{AK_pparams}}{

A list of 1 named vector and 1 matrix.

\describe{
\item{shared}{Only contains the value for \code{mu}, which was treated as fixed}
\item{specific}{Matrix of unit-specific parameter estimates. Rows are
parameter names, columns are unit names.}
}
}
}
\usage{
AK_pparams
}
\description{
Parameters for \code{AK_model()} in \code{pparams()} form
}
\keyword{datasets}
