% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1NB_benchmark.R
\name{ar1NB_benchmark}
\alias{ar1NB_benchmark}
\title{Produce an AR(1) negative binomial benchmark for panelPomp models}
\usage{
ar1NB_benchmark(
  x,
  init = NULL,
  start_vals = c(10, 0.5, 1),
  transform = TRUE,
  a = 2,
  ...
)
}
\arguments{
\item{x}{A panelPomp object or a matrix of data where each row is a unit and
each column is a time.}

\item{init}{Initial value \eqn{Y_0}. If \code{NULL}, uses \eqn{Y_0 = Y_1}.}

\item{start_vals}{Starting values for the parameter estimates.}

\item{transform}{Boolean, whether or not to use variable transformations in
the optimization.}

\item{a}{Only used if \code{transform == TRUE}. \code{a} is a numeric representing the
maximum value for the AR1 coefficient, and \code{-a} is the minimum value.}

\item{...}{other parameters to be passed into the \code{optim} function}
}
\value{
A list with a named vector of estimate parameters, unit log
likelihoods, a vector containing the total log likelihood, and a matrix of
conditional log likelihoods.
}
\description{
Produce an AR(1) negative binomial benchmark for panelPomp models
}
