% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma2_benchmark.R
\name{arma2_benchmark}
\alias{arma2_benchmark}
\title{Produce log-ARMA log-likelihood benchmark for panelPomp models}
\usage{
arma2_benchmark(x, order = c(2, 0, 1), na.rm = FALSE)
}
\arguments{
\item{x}{A panelPomp object or a matrix of data where each row is a unit and
each column is a time.}

\item{order}{A triple (p, d, q) for the ARIMA model fitted to the data.
Function code assumes that d will be set to 0.}

\item{na.rm}{Should NA's be removed when summing conditional log likelihoods
to obtain unit log likelihoods and total log likelihoods?}
}
\value{
A list with a named vector of unit log likelihoods, a vector
containing the total log likelihood, and a matrix of conditional log
likelihoods.
}
\description{
Produce log-ARMA log-likelihood benchmark for panelPomp models
}
