% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_benchmark2.R
\name{arma_benchmark2}
\alias{arma_benchmark2}
\title{Produce log-ARMA log-likelihood benchmark for panelPomp and spatPomp models}
\usage{
arma_benchmark2(xpomp, order = c(2, 0, 1))
}
\arguments{
\item{xpomp}{A panelPomp or spatPomp object.}

\item{order}{A triple (p, d, q) for the ARIMA model fitted to the data.
Function code assumes that d will be set to 0.}
}
\description{
Produce log-ARMA log-likelihood benchmark for panelPomp and spatPomp models
}
