% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_geo_cooling.R
\name{calc_geo_cooling}
\alias{calc_geo_cooling}
\title{Calculate amount by which rw.sd is cooled in geometric cooling scheme.}
\usage{
calc_geo_cooling(cooling_fraction_50, n = 1, m = 1, N = 730)
}
\arguments{
\item{cooling_fraction_50}{Fraction which rw.sd is multiplied by after 50
iterations.}

\item{n}{Time step number we want the cumulative cooling fraction for.}

\item{m}{Iteration number we want the cumulative cooling fraction for.}

\item{N}{Length of time series.}
}
\value{
A numeric vector of length one.
}
\description{
Calculate amount by which rw.sd is cooled in geometric cooling scheme.
}
\examples{
calc_geo_cooling(1/2, 1, m = 51, 730)
}
