% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_units.R
\name{choose_units}
\alias{choose_units}
\title{Choose which units to use from data}
\usage{
choose_units(data, units)
}
\arguments{
\item{data}{List of data in the form of \code{twentycities}.}

\item{units}{Character vector of unit names.}
}
\value{
List of data in the form of \code{twentycities} containing only
observations associated with \code{units}.
}
\description{
Choose which units to use from data
}
\examples{
choose_units(ur_measles, c("London", "Hastings", "Lees"))
}
