% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_twentycities.R
\name{clean_twentycities}
\alias{clean_twentycities}
\title{Perform minor cleanup on \code{twentycities} data}
\usage{
clean_twentycities()
}
\value{
Returns \code{twentycities} but with 3 observations set to NA.
}
\description{
Perform minor cleanup on \code{twentycities} data
}
\examples{
clean_twentycities()
}
