% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ur_measles.R
\name{clean_ur_measles}
\alias{clean_ur_measles}
\title{Clean the \code{ur_measles} data}
\usage{
clean_ur_measles()
}
\value{
Returns \code{ur_measles} but with tweaked observations.
}
\description{
Clean the \code{ur_measles} data
}
\examples{
clean_ur_measles()
}
