% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_param_guess_list.R
\name{construct_param_guess_list}
\alias{construct_param_guess_list}
\title{Title}
\usage{
construct_param_guess_list(
  shared_box_specs,
  specific_box_specs,
  nseq,
  buffer = 5e-08,
  seed = NULL,
  pos_params = c("R0", "mu", "rho", "sigmaSE"),
  unit_interval_params = c("cohort", "amplitude", "S_0", "E_0", "I_0", "R_0")
)
}
\arguments{
\item{unit_interval_params}{}
}
\description{
Title
}
