% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_specific_box_specs.R
\name{construct_specific_box_specs}
\alias{construct_specific_box_specs}
\title{Title}
\usage{
construct_specific_box_specs(specific_pparams, radii_tbl)
}
\arguments{
\item{radii_tbl}{}
}
\description{
Title
}
