% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_specific_pparams.R
\name{construct_specific_pparams}
\alias{construct_specific_pparams}
\title{Title}
\usage{
construct_specific_pparams(pparams_df, coef_choices)
}
\arguments{
\item{coef_choices}{}
}
\description{
Title
}
