% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_mp_file.R
\name{copy_mp_file}
\alias{copy_mp_file}
\title{Copy file from measlespkg misc directory}
\usage{
copy_mp_file(file_name = "generalPanelFit.R", to_path = ".")
}
\arguments{
\item{file_name}{String specifying name of file to copy}

\item{to_path}{String specifying where to copy the file to}
}
\value{
Returns NULL.
}
\description{
Copy file from measlespkg misc directory
}
