% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_logLik.ibpfd_spatPomp.R
\name{eval_logLik.ibpfd_spatPomp}
\alias{eval_logLik.ibpfd_spatPomp}
\title{Evaluate log likelihood of spatPomp model using the block particle filter}
\usage{
eval_logLik.ibpfd_spatPomp(
  model_obj_list,
  block_size = 1,
  ncores,
  np_pf,
  nreps,
  seed = NULL,
  divisor = NULL,
  return_pfilter_obj = FALSE
)
}
\arguments{
\item{model_obj_list}{List of model objects to evaluate log likelihood of.}

\item{block_size}{The number of spatial units per block.}

\item{ncores}{Number of cores to use for parallel computing.}

\item{np_pf}{Number of particles to use.}

\item{nreps}{Number of block particle filter repetitions.}

\item{seed}{Seed for particle filter. If NULL, does not set new seed.}

\item{divisor}{seed mod divisor*i is used to obtain seed for ith repetition.
If NULL, does not set new seed.}

\item{return_pfilter_obj}{Should the returned list include the bpfilter
object from the last repetition?}
}
\value{
A list containing the results of the likelihood evaluation.
}
\description{
Evaluate log likelihood of spatPomp model using the block particle filter
}
