% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_logLik.R
\name{eval_logLik}
\alias{eval_logLik}
\title{Evaluate log likelihood of model using particle filter, then save output.}
\usage{
eval_logLik(
  model_obj,
  write_to,
  ncores,
  np_pf,
  nreps,
  seed = NULL,
  divisor = NULL,
  pfilter_obj_write_to = NULL,
  print_time = TRUE
)
}
\arguments{
\item{model_obj}{Model object to evaluate log likelihood of.}

\item{write_to}{path/file to write output to.}

\item{ncores}{Number of cores to use for parallel computing.}

\item{np_pf}{Number of particles to use.}

\item{nreps}{Number of particle filter repetitions.}

\item{seed}{Seed for particle filter. If NULL, does not set new seed.}

\item{divisor}{seed mod divisor*i is used to obtain seed for ith repetition.
If NULL, does not set new seed.}

\item{pfilter_obj_write_to}{Where to save pfilter object from first
repetition. If NULL, pfilter object is not saved.}

\item{print_time}{Should time to run pfilter be printed?}
}
\value{
Nothing is returned, but several files are saved.
}
\description{
Evaluate log likelihood of model using particle filter, then save output.
}
