% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_results.R
\name{gather_results}
\alias{gather_results}
\title{Gather information on fit results from subdirectories and tidy them}
\usage{
gather_results(file_names, parent_dir = ".")
}
\arguments{
\item{file_names}{Character vector of names of rds files to load.}

\item{parent_dir}{Path to the directory that we want to search the
subdirectories of.}
}
\value{
\code{data.frame} where each row contains information for a unit from a
model replication. Columns include information such as the path to the Rds
file, the number of particles used for the fitting algorithm and likelihood
evaluation, and the parameters that the likelihood was evaluated at.
}
\description{
Given a parent directory and a vector of rds file names, this function will
attempt to load any rds files with those names under the parent directory. If
the rds file contains a \code{fit_results} object or an \code{EL_list} object, it will
call \code{tidy_results()} on it and row bind the resultings data frames. If an
error occurs at any point when trying to load and tidy a given file, the file
will be skipped over.
}
