% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_top_fits.R
\name{grab_top_fits}
\alias{grab_top_fits}
\title{Grab rows for top \code{n} fits based on log likelihood}
\usage{
grab_top_fits(x, top_n = 1, se_penalty = 0)
}
\arguments{
\item{x}{Object of class \code{EL_list}.}

\item{top_n}{Number of rows to grab.}

\item{se_penalty}{Penalizes log likelihood in ranking based on
\code{se*se_penalty}.}
}
\value{
Object of class \code{EL_list}.
}
\description{
Grab rows for top \code{n} fits based on log likelihood
}
\examples{
\dontrun{
AK_mod = AK_model()
EL_out = eval_logLik(
  list(AK_mod, AK_mod, AK_mod),
  ncores = 1,
  np_pf = 1,
  nreps = 2
)
grab_top_fits(EL_out)$fits[, 1:4]
}
}
