% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_top_unit_params.R
\name{grab_top_unit_params}
\alias{grab_top_unit_params}
\title{Title}
\usage{
grab_top_unit_params(pf_logLik_frame, pf_unitlogLik_frame, top_n = 1)
}
\arguments{
\item{top_n}{}
}
\description{
Title
}
