% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rw_sd.R
\name{make_rw_sd}
\alias{make_rw_sd}
\title{Turn named numeric vector into \code{rw_sd}}
\usage{
make_rw_sd(rw_sd_vec)
}
\arguments{
\item{rw_sd_vec}{Named numerical vector. Names are parameters, values are
desired random walk standard deviations. Names of initial value parameters
should end in _0.}
}
\value{
Returns \code{rw_sd} object.
}
\description{
Turn named numeric vector into \code{rw_sd}
}
\examples{
rw_sd_vec = c(alpha = 0.02, gamma = 0.01, E_0 = 0.003, I_0 = 0.006)
make_rw_sd(rw_sd_vec)
}
