% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sim_tbl.R
\name{make_sim_tbl}
\alias{make_sim_tbl}
\title{Make a tibble filled with simulated observations.}
\usage{
make_sim_tbl(sim_model, n_sims, true_model = NULL)
}
\arguments{
\item{sim_model}{\code{panelPomp} or \code{spatPomp} object to simulate from.}

\item{n_sims}{Number of simulations to perform.}

\item{true_model}{Object containing the real observations and times. Set to
\code{NULL} to exclude.}
}
\value{
A tibble with columns for observation time, replication name, unit,
and cases.
}
\description{
Make a tibble filled with simulated observations.
}
\examples{
\dontrun{
AK_mod = AK_model()
make_sim_tbl(AK_mod, n_sims = 10, true_model = AK_mod)
}
}
