% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spatMeaslesPomp.R
\name{make_spatMeaslesPomp}
\alias{make_spatMeaslesPomp}
\title{Make a spatPomp model using measles data}
\usage{
make_spatMeaslesPomp(
  data,
  model,
  interp_method = c("shifted_splines", "linear"),
  dt = 1/365.25
)
}
\arguments{
\item{data}{List in the format of \code{twentycities}.}

\item{model}{List of objects in the format of the output for the
\code{model_mechanics_XXX()} functions.}

\item{interp_method}{Method used to interpolate population and births.
Possible options are \code{"shifted_splines"} and \verb{"linear}.}

\item{dt}{Size of the time step.}
}
\value{
A panelPomp object using the data and model supplied.
}
\description{
Make a spatPomp model using measles data
}
