% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_sample_pparams.R
\name{mass_sample_pparams}
\alias{mass_sample_pparams}
\title{Quickly evaluate log likelihood at parameter sets and return best sets.}
\usage{
mass_sample_pparams(
  model_obj,
  pparams_list,
  Np = 1,
  Nt = 1,
  rank_scheme = c("total_ll", "ull"),
  top_n = 1,
  ncores = 1
)
}
\arguments{
\item{model_obj}{panelPomp object to run \code{pfilter()} on.}

\item{pparams_list}{List of parameters in the format of \code{pparams()} to
evaluate the log likelihood of \code{model_obj} at.}

\item{Np}{Number of particles to use.}

\item{Nt}{Number of time points to use.}

\item{rank_scheme}{How to rank the parameter sets.}

\item{top_n}{Number of best parameter sets to output.}

\item{ncores}{Number of cores to use.}
}
\value{
A list of parameters in the format of \code{pparams()}.
}
\description{
Quickly evaluate log likelihood at parameter sets and return best sets.
}
