% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_ppomp.R
\name{measles_ppomp}
\alias{measles_ppomp}
\title{He10 POMP model for UK measles.}
\usage{
measles_ppomp(
  data,
  starting_pparams,
  shared_str = "mu",
  sim_model = NULL,
  AK_interp = FALSE
)
}
\arguments{
\item{AK_interp}{}
}
\description{
He10 POMP model for UK measles.
}
