% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_ppomp_a2.R
\name{measles_ppomp_a2}
\alias{measles_ppomp_a2}
\title{Title}
\usage{
measles_ppomp_a2(
  use_nbinom = FALSE,
  shared_str = "mu",
  sim_model = NULL,
  AK_interp = FALSE
)
}
\arguments{
\item{AK_interp}{}
}
\description{
Title
}
