% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_ppomp_lmp_g.R
\name{measles_ppomp_lmp_g}
\alias{measles_ppomp_lmp_g}
\title{panelPOMP model for UK measles with linear model parameterization for
gamma}
\usage{
measles_ppomp_lmp_g(
  use_nbinom = FALSE,
  shared_str = "mu",
  reduced_model = FALSE,
  sim_model = NULL,
  AK_interp = FALSE
)
}
\arguments{
\item{AK_interp}{}
}
\description{
panelPOMP model for UK measles with linear model parameterization for
gamma
}
