% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_ppomp_lmp_g_t.R
\name{measles_ppomp_lmp_g_t}
\alias{measles_ppomp_lmp_g_t}
\title{panelPOMP model for UK measles with linear model parameterization for
gamma. Input population for linear model changes with time.}
\usage{
measles_ppomp_lmp_g_t(
  use_nbinom = FALSE,
  shared_str = "mu",
  reduced_model = FALSE,
  sim_model = NULL,
  AK_interp = FALSE
)
}
\arguments{
\item{AK_interp}{}
}
\description{
panelPOMP model for UK measles with linear model parameterization for
gamma. Input population for linear model changes with time.
}
