% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_ppomp_lmp_gp.R
\name{measles_ppomp_lmp_gp}
\alias{measles_ppomp_lmp_gp}
\title{panelPOMP model for UK measles with linear model parameterization for
gamma and psi.}
\usage{
measles_ppomp_lmp_gp(
  use_nbinom = FALSE,
  shared_str = "mu",
  reduced_model = FALSE,
  sim_model = NULL,
  AK_interp = FALSE
)
}
\arguments{
\item{AK_interp}{}
}
\description{
panelPOMP model for UK measles with linear model parameterization for
gamma and psi.
}
