% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles_ppomp_skel.R
\name{measles_ppomp_skel}
\alias{measles_ppomp_skel}
\title{Skeleton for measles_ppomp functions.}
\usage{
measles_ppomp_skel(
  rproc,
  dmeas,
  rmeas,
  extra_param_start,
  pt,
  paramnames,
  AK_interp,
  shared_str = "mu",
  sim_model = NULL,
  alternate_data = NULL
)
}
\arguments{
\item{alternate_data}{}
}
\description{
Skeleton for measles_ppomp functions.
}
