% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measlespkg-package.R
\docType{package}
\name{measlespkg-package}
\alias{measlespkg}
\alias{measlespkg-package}
\title{measlespkg: Fit POMP Models to Measles Data}
\description{
A collection of functions to make fitting POMP models to UK measles data easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AJAbkemeier/measlespkg}
  \item Report bugs at \url{https://github.com/AJAbkemeier/measlespkg/issues}
}

}
\author{
\strong{Maintainer}: Aaron Abkemeier \email{aaronabk@umich.edu} (\href{https://orcid.org/0009-0007-1936-2246}{ORCID})

}
\keyword{internal}
