% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mif2_alt.R
\name{mif2_alt}
\alias{mif2_alt}
\title{Title}
\usage{
mif2_alt(
  measles_ppomp_mod,
  param_guess_list,
  rs0,
  run_level,
  nmif,
  n_alt,
  nreps_mif2,
  np_mif2,
  cooling_frac,
  block = FALSE,
  shared_first = TRUE,
  seed = NULL,
  write_mif2_to
)
}
\arguments{
\item{write_mif2_to}{}
}
\description{
Title
}
