% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_001.R
\name{model_mechanics_001}
\alias{model_mechanics_001}
\title{He10 POMP model for UK measles}
\usage{
model_mechanics_001()
}
\value{
List of objects required for \code{make_measlesPomp}.
}
\description{
He10 POMP model for UK measles
}
\examples{
make_measlesPomp(twentycities, AK_pparams, model_mechanics_001())
}
