% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_001a.R
\name{model_mechanics_001a}
\alias{model_mechanics_001a}
\title{Experimental He10 POMP model for UK measles}
\usage{
model_mechanics_001a(shared_params = "mu")
}
\arguments{
\item{shared_params}{Character vector of parameters to be treated as shared.}
}
\value{
List of objects required for \code{make_measlesPomp}.
}
\description{
This model lacks alpha, and amplitude does not kick in during the Christmas
break.
}
