% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_001w.R
\name{model_mechanics_001w}
\alias{model_mechanics_001w}
\title{He10 POMP model for UK measles with an auxiliary weight parameter}
\usage{
model_mechanics_001w(param, U, shared_params = "mu", reg = 1)
}
\arguments{
\item{param}{Name of the parameter you would like to express in terms of a
shared parameter and a weighted unit-specific parameter.}

\item{U}{Number of units in the model.}

\item{shared_params}{Character vector of parameters to be treated as shared.}

\item{reg}{Number between 0 and 1 determining the strength of the
reparameterization constraint imposed at each observation step. Setting
this equal to 1 enforces the constraint exactly.}
}
\value{
List of objects required for \code{make_measlesPomp}.
}
\description{
He10 POMP model for UK measles with an auxiliary weight parameter
}
