% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_002.R
\name{model_mechanics_002}
\alias{model_mechanics_002}
\title{panelPOMP model with log-log relationship between gamma and standardized
1950 population}
\usage{
model_mechanics_002()
}
\value{
List of objects required for \code{make_measlesPomp()}.
}
\description{
panelPOMP model with log-log relationship between gamma and standardized
1950 population
}
