% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_004.R
\name{model_mechanics_004}
\alias{model_mechanics_004}
\title{panelPOMP model with log-log relationship for between gamma and
the standardized 1950 population, as well as between iota and the
standardized 1950 population}
\usage{
model_mechanics_004()
}
\value{
List of objects required for \code{make_measlesPomp()}.
}
\description{
panelPOMP model with log-log relationship for between gamma and
the standardized 1950 population, as well as between iota and the
standardized 1950 population
}
\examples{
new_pparams = AK_pparams
new_pparams$shared = c(
  new_pparams$shared,
  gamma1 = -0.6369,
  gamma0 = 4.6121,
  iota1 = 1.4172,
  iota0 = -1.9611
)
make_measlesPomp(twentycities, new_pparams, model_mechanics_004())
}
