% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_007 2.R,
%   R/model_mechanics_007.R
\name{model_mechanics_007}
\alias{model_mechanics_007}
\title{Experimental POMP model for UK measles}
\usage{
model_mechanics_007(U)

model_mechanics_007(U)
}
\value{
List of objects required for \code{make_measlesPomp}.

List of objects required for \code{make_spatMeaslesPomp}.
}
\description{
Beta is not a function of gamma or mu in this model.
Note: R0 loses its usual meaning in this model.
}
\examples{
make_measlesPomp(twentycities, AK_pparams, model_mechanics_007())
}
