% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_008.R
\name{model_mechanics_008}
\alias{model_mechanics_008}
\title{He10 POMP model for UK measles with AR(1) gamma noise}
\usage{
model_mechanics_008(shared_params = "mu")
}
\arguments{
\item{shared_params}{Character vector of parameters to be treated as shared.}
}
\value{
List of objects required for \code{make_measlesPomp}.
}
\description{
He10 POMP model for UK measles with AR(1) gamma noise
}
