% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_009.R
\name{model_mechanics_009}
\alias{model_mechanics_009}
\title{panelPOMP model with log-log relationship between iota and
the standardized 1950 population, as well as between psi and the
standardized 1950 population}
\usage{
model_mechanics_009(shared_params = "mu")
}
\arguments{
\item{shared_params}{Character vector of parameters to be treated as shared.}
}
\value{
List of objects required for \code{make_measlesPomp()}.
}
\description{
panelPOMP model with log-log relationship between iota and
the standardized 1950 population, as well as between psi and the
standardized 1950 population
}
