% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_010.R
\name{model_mechanics_010}
\alias{model_mechanics_010}
\title{panelPOMP model with log-log relationship between psi and the
standardized 1950 population}
\usage{
model_mechanics_010(shared_params = "mu")
}
\arguments{
\item{shared_params}{Character vector of parameters to be treated as shared.}
}
\value{
List of objects required for \code{make_measlesPomp()}.
}
\description{
panelPOMP model with log-log relationship between psi and the
standardized 1950 population
}
