% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_011se.R
\name{model_mechanics_011se}
\alias{model_mechanics_011se}
\title{panelPOMP model with log-log relationship between noisy iota and the
standardized 1950 population}
\usage{
model_mechanics_011se(shared_params = "mu")
}
\arguments{
\item{shared_params}{Character vector of parameters to be treated as shared
in addition to those hard-coded as shared.}
}
\value{
List of objects required for \code{make_measlesPomp()}.
}
\description{
panelPOMP model with log-log relationship between noisy iota and the
standardized 1950 population
}
