% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_fit_results.R
\name{new_fit_results}
\alias{new_fit_results}
\title{Make new fit_results object}
\usage{
new_fit_results(fitr_out, EL_out)
}
\arguments{
\item{fitr_out}{List of \code{mif2d.ppomp} or \code{ibpfd_spatPomp} objects.}

\item{EL_out}{\code{EL_list} object.}
}
\value{
List of class \code{fit_results} with first entry \code{fitr_out} and second
entry \code{EL_out}. \code{fitr_out} is a list of \code{mif2d.ppomp} or \code{ibpfd_spatPomp}
objects, and \code{EL_out} is an \code{EL_list}.
}
\description{
Make new fit_results object
}
