% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2.R
\name{obs2}
\alias{obs2}
\title{Extract the data matrix from a pomp/panelPomp/spatPomp object.}
\usage{
obs2(x)
}
\arguments{
\item{x}{An object of class \code{pomp}, \code{panelPomp}, or \code{spatPomp}.}
}
\value{
A matrix of observations with rows corresponding to units and columns
corresponding to time points.
}
\description{
Extract the data matrix from a pomp/panelPomp/spatPomp object.
}
\examples{
\dontrun{
obs2(AK_model())
}
}
