% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_mechanics.R
\name{panel_mechanics}
\alias{panel_mechanics}
\title{Make a list containing necessary specifications for a panelPomp model}
\usage{
panel_mechanics(rproc, dmeas, rmeas, rinit, pt, shared_params, specific_params)
}
\arguments{
\item{rproc}{Csnippet that simulates process step.}

\item{dmeas}{Csnippet that calculates conditional log likelihood.}

\item{rmeas}{Csnippet that simulates measurement.}

\item{rinit}{Csnippet that simulates initial states.}

\item{pt}{Csnippet specifiying how to transform parameters onto the
estimation scale.}

\item{shared_params}{Character vector of shared parameter names.}

\item{specific_params}{Character vector of unit-specific parameter names.}
}
\value{
The arguments in list form with class \code{panel_mechanics}.
}
\description{
Make a list containing necessary specifications for a panelPomp model
}
