% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traces.R
\name{plot_traces}
\alias{plot_traces}
\title{Title}
\usage{
plot_traces(
  mif2_alt_list,
  plot_shared = ".ALL",
  plot_specific = ".ALL",
  print_plots = TRUE,
  save_dir = NULL,
  width = 16,
  height = 10
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
