% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traces2.R
\name{plot_traces2}
\alias{plot_traces2}
\title{Plot traces from fitted objects}
\usage{
plot_traces2(fitr_list, plotted_params = ".ALL", log_y = FALSE, skip_n = 0)
}
\arguments{
\item{fitr_list}{List of \code{mif2d.ppomp} or \code{ibpfd_spatPomp} objects.}

\item{plotted_params}{String specifying which traces should be plotted. This
includes log-likelihoods and shared parameters. Set to \code{".ALL"} to plot all
traces.}

\item{log_y}{Boolean specifying whether y-axis should be scaled by log base
10 for positive-valued traces.}

\item{skip_n}{Skip plotting the first \code{skip_n} iterations. Useful when
extreme early values make it difficult to observe later behavior.}
}
\value{
Returns \code{NULL}.
}
\description{
Plot traces from fitted objects
}
