% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pparams_to_spatCoef.R
\name{pparams_to_spatCoef}
\alias{pparams_to_spatCoef}
\title{Convert list of \code{pparams} into \code{coef.spatPomp} format}
\usage{
pparams_to_spatCoef(pparams)
}
\arguments{
\item{pparams}{List of parameters in format of \code{pparams}.}
}
\value{
Named vector of parameters in format used by \code{coef.spatPomp}.
}
\description{
Convert list of \code{pparams} into \code{coef.spatPomp} format
}
