% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_round.panelPomp.R
\name{run_round.panelPomp}
\alias{run_round.panelPomp}
\title{Perform one round of model fitting}
\usage{
run_round.panelPomp(
  x,
  write_results_to,
  ncores,
  np_mif2,
  cooling_frac,
  rw_sd,
  nmif,
  block = FALSE,
  np_eval,
  nreps_eval,
  print_times = FALSE,
  initial_pparams_list,
  ...
)
}
\arguments{
\item{x}{\code{panelPomp} object.}

\item{write_results_to}{File path to save Rds file containing results to.}

\item{ncores}{Number of cores to use.}

\item{cooling_frac}{Cooling fraction to use when running fitting algorithm.}

\item{np_eval}{Number of particles to use when running \code{eval_logLik()}.}

\item{nreps_eval}{Number of replications to use when running \code{eval_logLik()}.}

\item{print_times}{Boolean for whether times to run the fitting algorithm and
\code{eval_logLik()} should be printed.}

\item{initial_pparams_list}{List of initial parameters in the format of
\code{pparams()}.}
}
\value{
Object of class \code{fit_results} containing a list of \code{mif2} objects and
a list of \code{EL_list} objects.
}
\description{
Perform one round of model fitting
}
