% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_round.R
\name{run_round}
\alias{run_round}
\title{Perform one round of model fitting}
\usage{
run_round(
  x,
  initial_pparams_list,
  write_results_to = NULL,
  ncores,
  np_fitr,
  cooling_frac,
  rw_sd_obj,
  N_fitr,
  panel_block = FALSE,
  np_eval,
  nreps_eval,
  print_times = FALSE,
  spat_block_size = 1,
  spat_sharedParNames = NULL,
  spat_unitParNames = NULL,
  spat_regression = 0.2,
  rparam = NULL
)
}
\arguments{
\item{x}{A \code{panelPomp} or \code{spatPomp} object.}

\item{initial_pparams_list}{List of initial parameters in the format of
\code{panelPomp::coef()} with \verb{format = "list}. Each entry in the list specifies
the initial parameters for one replication of the fitting algorithm.}

\item{write_results_to}{File path to save Rds file containing results to.}

\item{ncores}{Number of cores to use.}

\item{np_fitr}{Number of particles to use when running the fitting algorithm.}

\item{cooling_frac}{Cooling fraction to use when running fitting algorithm.}

\item{rw_sd_obj}{Object of class \code{rw_sd} specifying random walk standard
deviations to use when running the fitting algorithm.}

\item{N_fitr}{Number of iterations to use when running the fitting algorithm.}

\item{panel_block}{Boolean specifying whether to perform block resampling of
specific parameters for a \code{panelPomp} model. Only used when \code{x} is a
\code{panelPomp} object.}

\item{np_eval}{Number of particles to use when running \code{eval_logLik()}.}

\item{nreps_eval}{Number of replications to use when running \code{eval_logLik()}.}

\item{print_times}{Boolean for whether times to run the fitting algorithm and
\code{eval_logLik()} should be printed.}

\item{spat_block_size}{Block size used when fitting and evaluating \code{spatPomp}
models. Only used when \code{x} is a \code{spatPomp} object.}

\item{spat_sharedParNames}{Character vector of parameters to be treated as
shared by \code{ibpf()}. Only used when \code{x} is a \code{spatPomp} object.}

\item{spat_unitParNames}{Character vector of parameters to be treated as
unit-specific by \code{ibpf()}. Only used when \code{x} is a \code{spatPomp} object.}

\item{spat_regression}{Numeric specifying the mean-regressing coefficient to
use when fitting shared parameters in a \code{spatPomp} model. Only used when
\code{x} is a \code{spatPomp} object.}

\item{rparam}{If using a reparameterization for GMCAP, specify which
parameter is reparameterized using this argument so the reparameterization
constraint can be applied before evaluating the log likelihood. Set to NULL
otherwise.}
}
\value{
Object of class \code{fit_results} containing a list of \code{mif2d.ppomp} or
\code{ibpfd_spatPomp} objects and a list of \code{EL_list} objects.
}
\description{
Depending on the model used, \code{run_round} will use either \link[panelPomp:mif2]{panelPomp::mif2} or
\link[spatPomp:ibpf]{spatPomp::ibpf} to fit it, then use \code{eval_logLik()} to estimate the log
likelihood of the fit. The results are saved using \link[pomp:bake]{pomp::bake}.
}
