% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_round.spatPomp.R
\name{run_round.spatPomp}
\alias{run_round.spatPomp}
\title{Perform one round of model fitting}
\usage{
\method{run_round}{spatPomp}(
  x,
  write_results_to,
  ncores,
  np_mif2,
  cooling_frac,
  rw_sd,
  nmif,
  block = TRUE,
  np_eval,
  nreps_eval,
  print_times = FALSE,
  initial_pparams_list,
  block_size = 1,
  spat_regression = 0.1,
  ...
)
}
\arguments{
\item{x}{\code{spatPomp} object.}

\item{write_results_to}{File path to save Rds file containing results to.}

\item{ncores}{Number of cores to use.}

\item{np_mif2}{Number of particles to use when running \code{mif2()}.}

\item{cooling_frac}{Cooling fraction to use when running \code{mif2()}.}

\item{rw_sd}{Random walk standard deviation to use when running \code{mif2()}.}

\item{nmif}{Number of iterations to use when running \code{mif2()}.}

\item{block}{Boolean specifying whether to perform block resampling of
specific parameters.}

\item{np_eval}{Number of particles to use when running \code{eval_logLik()}.}

\item{nreps_eval}{Number of replications to use when running \code{eval_logLik()}.}

\item{print_times}{Boolean for whether times to run \code{mif2()} and
\code{eval_logLik()} should be printed.}

\item{initial_pparams_list}{List of initial parameters in the format of
\code{pparams()}.}

\item{block_size}{The number of spatial units per block.}

\item{spat_regression}{Fraction of each extended shared parameter regressed
toward the unit mean. Not required when all parameters are unit-specific.}

\item{...}{Additional arguments for other methods.}

\item{sharedParNames}{Estimated parameters that are shared between units.}

\item{unitParParNames}{Estimated parameters that are unit-specific.}
}
\value{
Object of class \code{fit_results} containing a list of \code{ibpfd_spatPomp}
objects and a list of \code{EL_list} objects.
}
\description{
Perform one round of model fitting
}
