% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_initial_pparams_rc.R
\name{sample_initial_pparams_rc}
\alias{sample_initial_pparams_rc}
\title{Use centers and radii to sample initial parameters from box}
\usage{
sample_initial_pparams_rc(
  shared_box_specs,
  specific_pparams_df,
  radii_tbl,
  n_draws,
  buffer = 5e-08,
  pos_params = c("R0", "mu", "rho", "sigmaSE"),
  unit_interval_params = c("cohort", "amplitude", "S_0", "E_0", "I_0", "R_0")
)
}
\arguments{
\item{shared_box_specs}{\code{tbl} with \code{param}, \code{center}, and \code{radius} columns.}

\item{specific_pparams_df}{\code{data.frame} of specific parameter values in
format of pparams.}

\item{radii_tbl}{\code{tbl} with \code{param} and \code{radius} columns.}

\item{n_draws}{Number of initial parameter sets to draw.}

\item{buffer}{The minimum space required between the boundaries of the
sampling box and the boundaries of the possible values for the parameters.}

\item{pos_params}{Character vector of parameters which must be positive.}

\item{unit_interval_params}{Character vector of parameters which must be
between 0 and 1.}
}
\value{
A list of parameters sets in the \code{pparams()} format.
}
\description{
Use centers and radii to sample initial parameters from box
}
