% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_initial_pparams_ul.R
\name{sample_initial_pparams_ul}
\alias{sample_initial_pparams_ul}
\title{Use upper and lower bounds to sample initial parameters from box}
\usage{
sample_initial_pparams_ul(shared_box_specs, specific_box_specs, units, n_draws)
}
\arguments{
\item{shared_box_specs}{\code{tbl} with \code{param}, \code{lower}, and \code{upper} columns.}

\item{specific_box_specs}{\code{tbl} with \code{param}, \code{lower}, and \code{upper} columns.
format of pparams.}

\item{units}{Character vector of unit names.}

\item{n_draws}{Number of initial parameter sets to draw.}
}
\value{
A list of parameters sets in the \code{pparams()} format.
}
\description{
Use upper and lower bounds to sample initial parameters from box
}
