% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_plots.R
\name{sim_plots}
\alias{sim_plots}
\title{Title}
\usage{
sim_plots(
  true_model,
  sim_model,
  n_sims = 100,
  print_plots = TRUE,
  save_dir = NULL,
  width = 16,
  height = 12
)
}
\arguments{
\item{height}{}
}
\description{
Title
}
