% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spat_mechanics.R
\name{spat_mechanics}
\alias{spat_mechanics}
\title{Make a list containing necessary specifications for a spatPomp model}
\usage{
spat_mechanics(
  rproc,
  dmeas,
  dunit_measure,
  rmeas,
  rinit,
  pt,
  shared_params,
  specific_params,
  U
)
}
\arguments{
\item{rproc}{spatPomp_Csnippet that simulates process step.}

\item{dmeas}{spatPomp_Csnippet that calculates conditional log likelihood.}

\item{dunit_measure}{spatPomp_Csnippet that calculates conditional log
likelihood for a single unit.}

\item{rmeas}{spatPomp_Csnippet that simulates measurement.}

\item{rinit}{spatPomp_Csnippet that simulates initial states.}

\item{pt}{Csnippet specifiying how to transform parameters onto the
estimation scale.}

\item{shared_params}{Character vector of shared parameter names.}

\item{specific_params}{Character vector of unit-specific parameter names.}

\item{U}{Number of units.}
}
\value{
The arguments in list form with class \code{spat_mechanics}.
}
\description{
Make a list containing necessary specifications for a spatPomp model
}
