% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatCoef_to_pparams.R
\name{spatCoef_to_pparams}
\alias{spatCoef_to_pparams}
\title{Convert \code{coef.spatPomp} to \code{pparams} format}
\usage{
spatCoef_to_pparams(spatCoef, units)
}
\arguments{
\item{spatCoef}{Named numeric vector in format of \code{coef.spatPomp} object.
This function assumes that parameters are named according to the scheme
"param1, param2, param3,..." where "param" is the base parameter name and
doesn't end with a digit other than 0. Hence, R0 and S_0 would be valid
base parameter names but gamma1 would not be.}

\item{units}{Character vector of unit names in the order they are enumerated
in \code{spatCoef}.}
}
\value{
Parameters in the format of \code{panelPomp::coef()} with
\code{format = "list"}.
}
\description{
Convert \code{coef.spatPomp} to \code{pparams} format
}
