% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pfilter_dfs.R
\name{tidy_pfilter_dfs}
\alias{tidy_pfilter_dfs}
\title{Create tidy data frame from saved pfilter output.}
\usage{
tidy_pfilter_dfs(LL_df, ULL_df)
}
\arguments{
\item{ULL_df}{}
}
\description{
Create tidy data frame from saved pfilter output.
}
