% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_results.R
\name{tidy_results}
\alias{tidy_results}
\alias{tidy_results.EL_list}
\alias{tidy_results.fit_results}
\title{Convert \code{fit_results} or \code{EL_list} output into tidy data frame}
\usage{
tidy_results(x, path = NA)

\method{tidy_results}{EL_list}(x, path = NA)

\method{tidy_results}{fit_results}(x, path = NA)
}
\arguments{
\item{x}{Object of type \code{fit_results} or \code{EL_list}.}

\item{path}{Character specifying where \code{x} was obtained from. Mostly exists
for use by \code{\link[=gather_results]{gather_results()}}. Very optional.}
}
\value{
\code{data.frame} where each row contains information for a unit from a
model replication. Columns include information such as the number of
particles used for the fitting algorithm and likelihood evaluation, and the
parameters that the likelihood was evaluated at.
}
\description{
Convert \code{fit_results} or \code{EL_list} output into tidy data frame
}
