% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_traces.R
\name{tidy_traces}
\alias{tidy_traces}
\title{Return tidy tibble of traces}
\usage{
tidy_traces(x)
}
\arguments{
\item{x}{A list of \code{mif2d.ppomp} objects.}
}
\value{
A tibble with columns for replication number, unit name, trace name,
and trace value. If the parameter in question is shared, the unit name will
show up as "SHARED".
}
\description{
Return tidy tibble of traces
}
