% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fit_results.R
\name{update_fit_results}
\alias{update_fit_results}
\title{Update structure of older versions of fit_results objects}
\usage{
update_fit_results(x)
}
\arguments{
\item{x}{\code{fit_results} object to be updated.}
}
\value{
Object of class \code{fit_results} with structure compatible with more
recent versions of this package
}
\description{
Update structure of older versions of fit_results objects
}
