% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ur_measles.R
\docType{data}
\name{ur_measles}
\alias{ur_measles}
\title{Weekly reported measles data for 1422 locations in the UK}
\format{
\subsection{\code{ur_measles}}{

A list of 3 tibbles. Unit names ending in \code{.RD} are for rural areas; other
unit names are for urban areas. NOTE: not all units have coordinates.
}

\subsection{\code{ur_measles$measles}:}{

\describe{
\item{unit}{City name}
\item{data}{Date of observation}
\item{cases}{Number of measles cases reported during the week}
}
}

\subsection{\code{ur_measles$demog}:}{

\describe{
\item{unit}{City name}
\item{year}{Year that demography was recorded}
\item{pop}{Population}
\item{births}{Births}
}
}

\subsection{\code{ur_measles$coord}:}{

\describe{
\item{unit}{City name}
\item{long}{Longitude of city}
\item{lat}{Latitude of city}
}
}
}
\usage{
ur_measles
}
\description{
Weekly reported measles data for 1422 locations in the UK
}
\keyword{datasets}
