% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK_model.R
\name{AK_model}
\alias{AK_model}
\title{Create panelPomp object using AK's re-fitting of He10's model.}
\usage{
AK_model()
}
\value{
panelPomp object.
}
\description{
AK's model is like He10's model EXCEPT that population and
births are interpolated using smooth splines and the interpolated covariates
are shifted +0.5 years. In contrast, He10 uses linear interpolation and
makes no shift. In addition, AK's MLE's are different.
}
\examples{
AK_model()
}
