% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_to_pparams.R
\name{coef_to_pparams}
\alias{coef_to_pparams}
\title{Convert coef-style object to pparams-style object.}
\usage{
coef_to_pparams(coef)
}
\arguments{
\item{coef}{Vector in the style of \code{coef(panelPomp_obj)}. That is, a numeric
vector with names styled as "\code{shared_parameter}" or
"\code{specific_parameter[unit]}".}
}
\value{
A list of length 2 in the style of \code{pparams(panelPomp_obj)}. That is,
a numeric vector with shared parameter names, and a matrix with specific
parameters as row names and units as column names.
}
\description{
Vector entries with name in the style of \code{param[unit]} are assumed to be
unit-specific whereas those with name in the style of \code{param} are assumed to
be shared.
}
\examples{
coef_out = panelPomp::coef(AK_model())
coef_to_pparams(coef_out)
}
