% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_top_fits.R
\name{combine_top_fits}
\alias{combine_top_fits}
\title{Combine parameter estimates with best unit log likelihoods}
\usage{
combine_top_fits(x, top_n = 1, se_penalty = 0)
}
\arguments{
\item{x}{Object of class \code{EL_list}.}

\item{top_n}{Number of rows to grab.}

\item{se_penalty}{Penalizes log likelihood in ranking based on
\code{se*se_penalty}.}
}
\value{
Object of class \code{EL_list}.
}
\description{
Combine parameter estimates with best unit log likelihoods
}
\examples{
AK_mod = AK_model()
EL_out = eval_logLik(
  list(AK_mod, AK_mod, AK_mod),
  ncores = 1,
  np_pf = 1,
  nreps = 2
)
EL_out$fits
combine_top_specific_fits(EL_out, top_n = 2)
}
