% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_top_pparams.R
\name{duplicate_top_pparams}
\alias{duplicate_top_pparams}
\title{Duplicate top fits from \code{EL_list}}
\usage{
duplicate_top_pparams(
  x,
  out_length,
  top_n = 1,
  combine = FALSE,
  units = NULL,
  is_spat = FALSE
)
}
\arguments{
\item{x}{Object of class \code{EL_list}.}

\item{out_length}{Number of parameter sets to output.}

\item{top_n}{Number of top fits to duplicate.}

\item{combine}{Boolean specifying whether best specific fits should be
combined.}

\item{units}{Character vector of unit names, which is necessary when
duplicating the parameters of a \code{spatPomp} object. If the parameters belong
to a \code{panelPomp} object, leave as \code{NULL}.}

\item{is_spat}{Indicates whether parameters in \code{x} are for a spatPOMP model.
Function will fail to return desired results if wrong.}
}
\value{
List of parameters in the form of \code{\link[panelPomp:pparams]{panelPomp::pparams()}}.
}
\description{
Duplicate top fits from \code{EL_list}
}
\examples{
AK_mod = AK_model()
EL_out = eval_logLik(
  list(AK_mod, AK_mod, AK_mod),
  ncores = 1,
  np_pf = 1,
  nreps = 2
)
duplicate_top_pparams(EL_out, out_length = 6, top_n = 2)
}
