% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_logLik.R
\name{eval_logLik}
\alias{eval_logLik}
\title{Evaluate the log likelihood of a model using \code{pfilter()}}
\usage{
eval_logLik(
  model_obj_list,
  block_size = 1,
  ncores,
  np_pf,
  nreps,
  seed = NULL,
  divisor = NULL,
  return_n_pfilter_objs = 0
)
}
\arguments{
\item{model_obj_list}{List of \code{panelPomp} or \code{spatPomp} objects to evaluate
the log likelihood of.}

\item{block_size}{The number of spatial units per block. Only used when
evaluating are \code{spatPomp} models. (NOTE: function will break for any block
size other than 1; this will be fixed when choosing other block sizes seems
worthwhile.)}

\item{ncores}{Number of cores to use for parallel computing.}

\item{np_pf}{Number of particles to use.}

\item{nreps}{Number of particle filter repetitions.}

\item{seed}{Seed for particle filter. If NULL, does not set new seed.}

\item{divisor}{\code{seed} mod \code{divisor*i} is used to obtain seed for ith
repetition. If NULL, does not set new seed.}

\item{return_n_pfilter_objs}{Number of \code{pfilter} objects to return. Currently
only works for \code{panelPomp} models.}
}
\value{
Object of type \code{EL_list}, a list of data frames containing log
likelihood and standard error estimates. The unit log likelihoods estimated
for \code{spatPomp} objects are not real unit log likelihoods given that the
units are dependent, but they may still be useful when evaluating model
fit.
}
\description{
Evaluate the log likelihood of a model using \code{pfilter()}
}
\examples{
model_list = list(AK_model())
eval_logLik(model_list, ncores = 1, np_pf = 3, nreps = 2)
}
