% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_logLik.R
\name{eval_logLik}
\alias{eval_logLik}
\title{Evaluate log likelihood of model using particle filter, then save output.}
\usage{
eval_logLik(
  model_obj_list,
  ncores,
  np_pf,
  nreps,
  seed = NULL,
  divisor = NULL,
  return_pfilter_obj = FALSE
)
}
\arguments{
\item{model_obj_list}{List of model objects to evaluate log likelihood of.}

\item{ncores}{Number of cores to use for parallel computing.}

\item{np_pf}{Number of particles to use.}

\item{nreps}{Number of particle filter repetitions.}

\item{seed}{Seed for particle filter. If NULL, does not set new seed.}

\item{divisor}{seed mod divisor*i is used to obtain seed for ith repetition.
If NULL, does not set new seed.}

\item{return_pfilter_obj}{Should the returned list include a pfilter object?
repetition. If NULL, pfilter object is not saved.}
}
\value{
A list of data frames. Each row of \code{fits} contains the estimated
total log likelihood, standard error, and parameters for
}
\description{
Evaluate log likelihood of model using particle filter, then save output.
}
\examples{
model_list = list(AK_model())
eval_logLik(model_list, ncores = 1, np_pf = 3, nreps = 2)
}
