% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_logLik.R
\name{eval_logLik}
\alias{eval_logLik}
\title{Evaluate the log likelihood of a model using \code{pfilter()}}
\usage{
eval_logLik(
  model_obj_list,
  block_size = 1,
  ncores,
  np_pf,
  nreps,
  seed = NULL,
  divisor = NULL,
  return_n_pfilter_objs = 0
)
}
\arguments{
\item{model_obj_list}{List of \code{panelPomp} or \code{spatPomp} objects to evaluate
the log likelihood of.}

\item{block_size}{The number of spatial units per block. Only used when
evaluating are \code{spatPomp} models. (NOTE: function will break for any block
size other than 1; this will be fixed when choosing other block sizes seems
worthwhile.)}

\item{ncores}{Number of cores to use for parallel computing.}

\item{np_pf}{Number of particles to use.}

\item{nreps}{Number of particle filter repetitions.}

\item{seed}{Seed for particle filter. If NULL, does not set new seed.}

\item{divisor}{\code{seed} mod \code{divisor*i} is used to obtain seed for ith
repetition. If NULL, does not set new seed.}

\item{return_n_pfilter_objs}{Number of \code{pfilter} objects to return.}
}
\value{
Object of type \code{EL_list}, a list of data frames containing log
likelihood and standard error estimates.

Note that the unit log likelihoods estimated for \code{spatPomp} objects are not
real unit log likelihoods given that the units are dependent, but they may
still be useful when evaluating model fit.

For \code{panelPomp}, log likelihood estimates are obtained by using
\code{\link[panelPomp:panel_logmeanexp]{panelPomp::panel_logmeanexp()}} over the \code{pfilter()} log likelihood
replications. For \code{spatPomp}, log likelihood estimates are obtained by
using \code{\link[pomp:logmeanexp]{pomp::logmeanexp()}} over the \code{pfilter()} log likelihood
replications. For both model types, unit log likelihood estimates are
obrained by using \code{logmeanexp()} over the unit log likelihood replications
for each unit, and conditional log likelihood estimates are obtained by
using \code{logmeanexp()} over the condtional log likelihood replications for
each unit and time point.
}
\description{
Evaluate the log likelihood of a model using \code{pfilter()}
}
\examples{
model_list = list(AK_model())
eval_logLik(model_list, ncores = 1, np_pf = 3, nreps = 2)
}
