% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_results.R
\name{gather_results}
\alias{gather_results}
\title{Gather information on fit results from subdirectories and tidy them}
\usage{
gather_results(file_names, parent_dir = ".")
}
\arguments{
\item{file_names}{Character vector of names of rds files to load.}

\item{parent_dir}{Path to the directory that we want to search the
subdirectories of.}
}
\value{
\code{data.frame} where each row contains information for a unit from a
\code{mif2} replication. Columns include information such as the path to the rds
file, the number of particles used for \code{mif2} and likelihood evaluation,
and the parameters that the likelihood was evaluated at.
}
\description{
Gather information on fit results from subdirectories and tidy them
}
