% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_measlesPomp.R
\name{make_measlesPomp}
\alias{make_measlesPomp}
\title{Skeleton for measles_ppomp functions}
\usage{
make_measlesPomp(
  data,
  starting_pparams,
  model,
  AK_interp = TRUE,
  sim_obs_list = NULL
)
}
\arguments{
\item{sim_obs_list}{}
}
\description{
Skeleton for measles_ppomp functions
}
\examples{
make_measlesPomp(twentycities, AK_pparams, model_mechanics_001())
}
