% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_002.R
\name{model_mechanics_002}
\alias{model_mechanics_002}
\title{panelPOMP model with log-log relationship between gamma and standardized
1950 population}
\usage{
model_mechanics_002()
}
\value{
List of objects required for \code{make_measlesPomp()}.
}
\description{
panelPOMP model with log-log relationship between gamma and standardized
1950 population
}
\examples{
new_pparams = AK_pparams
new_pparams$shared = c(new_pparams$shared, gamma1 = -0.6369, gamma0 = 4.6121)
make_measlesPomp(twentycities, new_pparams, model_mechanics_002())
}
