% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_005.R
\name{model_mechanics_005}
\alias{model_mechanics_005}
\title{panelPOMP model with log-log relationship for between gamma and
the standardized 1950 population, as well as between psi and the
standardized 1950 population}
\usage{
model_mechanics_005()
}
\value{
List of objects required for \code{make_measlesPomp()}.
}
\description{
panelPOMP model with log-log relationship for between gamma and
the standardized 1950 population, as well as between psi and the
standardized 1950 population
}
