% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mechanics_007.R
\name{model_mechanics_007}
\alias{model_mechanics_007}
\title{spatPomp model for UK measles}
\usage{
model_mechanics_007(U, shared_params = "muD")
}
\arguments{
\item{U}{Number of units to be included.}

\item{shared_params}{Character vector of parameters to be treated as shared.}
}
\value{
List of objects required for \code{make_spatMeaslesPomp}.
}
\description{
spatPomp model for UK measles
}
