% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traces.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot traces from \code{mif2} objects}
\usage{
plot_traces(
  mif2_obj,
  plot_shared = ".ALL",
  plot_specific = ".ALL",
  print_plots = TRUE,
  save_dir = NULL,
  width = 16,
  height = 10
)
}
\arguments{
\item{mif2_obj}{\code{mif2} object.}

\item{plot_shared}{String specifying which shared value traces should be
plotted. This includes log-likelihoods and shared parameters. Set to \code{".ALL"}
to plot all shared value traces.}

\item{plot_specific}{String specifying which unit-specific value traces
should be plotted. This includes unit log-likelihoods and specific
parameters. Set to \code{".ALL"} to plot all unit-specific value traces.}

\item{print_plots}{Boolean specifying whether plots should be printed.}

\item{save_dir}{String specifying where plots should be saved. No plots are
saved if \code{NULL}.}

\item{width}{Width of plots.}

\item{height}{Height of plots.}
}
\value{
Returns \code{NULL}
}
\description{
Plot traces from \code{mif2} objects
}
\examples{
AK_mod = AK_model()
DEFAULT_SD = 0.02
IVP_DEFAULT_SD = DEFAULT_SD*3
INITIAL_RW_SD = c(
  S_0 = IVP_DEFAULT_SD,
  E_0 = IVP_DEFAULT_SD,
  I_0 = IVP_DEFAULT_SD,
  R_0 = IVP_DEFAULT_SD,
  R0 = DEFAULT_SD,
  sigmaSE = DEFAULT_SD,
  amplitude = DEFAULT_SD,
  rho = DEFAULT_SD,
  gamma = DEFAULT_SD,
  psi = DEFAULT_SD,
  iota = DEFAULT_SD,
  sigma = DEFAULT_SD,
  cohort = DEFAULT_SD,
  alpha = DEFAULT_SD*10^(-2),
  mu = 0
)
mif2_out = panelPomp::mif2(
  AK_mod,
  Np = 1,
  cooling.fraction.50 = 0.5,
  rw.sd = make_rw_sd(INITIAL_RW_SD),
  cooling.type = "geometric",
  Nmif = 3,
  shared.start = pparams(AK_mod)$shared,
  specific.start = pparams(AK_mod)$specific
)
plot_traces(mif2_out)
}
