% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traces.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot traces from fitted objects}
\usage{
plot_traces(
  fitr_list,
  plot_shared = ".ALL",
  plot_specific = ".ALL",
  print_plots = TRUE,
  save_dir = NULL,
  width = 16,
  height = 10,
  log_y = FALSE
)
}
\arguments{
\item{fitr_list}{List of \code{mif2d.ppomp} or \code{ibpfd_spatPomp} objects.}

\item{plot_shared}{String specifying which shared value traces should be
plotted. This includes log-likelihoods and shared parameters. Set to \code{".ALL"}
to plot all shared value traces.}

\item{plot_specific}{String specifying which unit-specific value traces
should be plotted. This includes unit log-likelihoods and specific
parameters. Set to \code{".ALL"} to plot all unit-specific value traces.}

\item{print_plots}{Boolean specifying whether plots should be printed.}

\item{save_dir}{String specifying where plots should be saved. No plots are
saved if \code{NULL}.}

\item{width}{Width of plots.}

\item{height}{Height of plots.}

\item{log_y}{Boolean specifying whether y-axis should be scaled by log base
10.}
}
\value{
Returns \code{NULL}.
}
\description{
Plot traces from fitted objects
}
