% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traces2.R
\name{plot_traces2}
\alias{plot_traces2}
\title{Plot traces from fitted objects}
\usage{
plot_traces2(
  traces_tbl,
  plotted_params = ".ALL",
  log_y = NULL,
  logit_y = NULL,
  skip_n = 0,
  lq = 0,
  pseudo_shared_param = NULL
)
}
\arguments{
\item{traces_tbl}{Data frame in format of \code{tidy_traces()} output.}

\item{plotted_params}{String specifying which traces should be plotted. This
includes log-likelihoods and shared parameters. Set to \code{".ALL"} to plot all
traces.}

\item{log_y}{Character vector specifying for which parameters traces the
y-axis should be log scaled.}

\item{logit_y}{Character vector specifying for which parameters traces the
y-axis should be logit scaled.}

\item{skip_n}{Skip plotting the first \code{skip_n} iterations. Useful when
extreme early values make it difficult to observe later behavior.}

\item{lq}{Lower quantile cutoff. When shading the traces with the log
likelihood of the fit, values below the lower quantile are rounded up to
the lower quantile.}

\item{pseudo_shared_param}{Name of a pseudo-shared parameter in the model
without the number attached at the end. Supplying this argument will cause
the function to to treat the pseudo-shared parameter as unit-specific. Does
nothing if set to NULL.}
}
\value{
Returns \code{NULL}.
}
\description{
Plot traces from fitted objects
}
