% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_round.R
\name{run_round}
\alias{run_round}
\title{Perform one round of model fitting}
\usage{
run_round(
  x,
  write_results_to,
  ncores,
  np_mif2,
  cooling_frac,
  rw_sd,
  nmif,
  block = FALSE,
  np_eval,
  nreps_eval,
  print_times = FALSE,
  ...
)
}
\arguments{
\item{x}{An object used to select a method. Should either be
\code{panelPomp} object or an object of class \code{fit_results}.}

\item{write_results_to}{File path to save Rds file containing results to.}

\item{ncores}{Number of cores to use.}

\item{np_mif2}{Number of particles to use when running \code{mif2()}.}

\item{cooling_frac}{Cooling fraction to use when running \code{mif2()}.}

\item{rw_sd}{Random walk standard deviation to use when running \code{mif2()}.}

\item{nmif}{Number of iterations to use when running \code{mif2()}.}

\item{block}{Boolean specifying whether to perform block resampling of
specific parameters.}

\item{np_eval}{Number of particles to use when running \code{eval_logLik()}.}

\item{nreps_eval}{Number of replications to use when running \code{eval_logLik()}.}

\item{print_times}{Boolean for whether times to run \code{mif2()} and
\code{eval_logLik()} should be printed.}

\item{...}{Additional arguments for other methods.}
}
\value{
Object of class \code{fit_results} containing a list of \code{mif2} objects
and a list of \code{EL_list} objects.
}
\description{
Perform one round of model fitting
}
