% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_initial_pparams_ul.R
\name{sample_initial_pparams_ul}
\alias{sample_initial_pparams_ul}
\title{Use upper and lower bounds to sample initial parameters from box}
\usage{
sample_initial_pparams_ul(sh_ul, sp_ul, units, n_draws)
}
\arguments{
\item{sh_ul}{\code{tbl} with \code{param}, \code{lower}, and \code{upper} columns.}

\item{sp_ul}{\code{tbl} with \code{param}, \code{lower}, and \code{upper} columns.
format of pparams.}

\item{units}{Character vector of unit names.}

\item{n_draws}{Number of initial parameter sets to draw.}
}
\value{
A list of parameters sets in the \code{pparams()} format.
}
\description{
Use upper and lower bounds to sample initial parameters from box
}
\examples{
\dontrun{
AK_mod = AK_model()
shared_bounds = tibble::tribble(
~param, ~lower,     ~upper,
"mu",        0.02,     0.02
)
specific_bounds = tibble::tribble(
  ~param,       ~lower,        ~upper,
  "R0",             10,            60,
  "rho",           0.1,           0.9,
  "sigmaSE",      0.04,           0.1,
  "amplitude",     0.1,           0.6,
  "S_0",          0.01,          0.07,
  "E_0",      0.000004,        0.0001,
  "I_0",      0.000003,         0.001,
  "R_0",           0.9,          0.99,
  "sigma",          25,           100,
  "iota",        0.004,             3,
  "psi",          0.05,             3,
  "alpha",       0.935,          1.05,
  "cohort",        0.1,           0.7,
  "gamma",          25,           320
)
sample_initial_pparams_ul(
  sh_ul = shared_bounds,
  sp_ul = specific_bounds,
  units = names(AK_mod),
  n_draws = 3
)
}
}
