% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_plots.R
\name{sim_plots}
\alias{sim_plots}
\title{Plot simulated observations against real observations}
\usage{
sim_plots(
  true_model,
  sim_model,
  n_sims = 100,
  print_plots = TRUE,
  save_dir = NULL,
  width = 16,
  height = 12
)
}
\arguments{
\item{true_model}{\code{panelPomp} object to get true observations from.}

\item{sim_model}{\code{panelPomp} object to simulate observations from.}

\item{n_sims}{Number of simulations to generate.}

\item{print_plots}{Boolean specifying whether plots should be printed.}

\item{save_dir}{String specifying where plots should be saved. No plots are
saved if \code{NULL}.}

\item{width}{Width of plots.}

\item{height}{Height of plots.}
}
\value{
Returns \code{NULL}.
}
\description{
Plot simulated observations against real observations
}
\examples{
\dontrun{
mod = AK_model()
sim_plots(mod, mod, n_sims = 10)
}
}
