% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatCoef_to_pparams.R
\name{spatCoef_to_pparams}
\alias{spatCoef_to_pparams}
\title{Convert \code{coef.spatPomp} to \code{pparams} format}
\usage{
spatCoef_to_pparams(spatCoef, units)
}
\arguments{
\item{spatCoef}{Named numeric vector in format of \code{coef.spatPomp} object.}

\item{units}{Character vector of unit names in the order they are enumerated
in \code{spatCoef}.}
}
\value{
Parameters in the format of \link[panelPomp:pparams]{panelPomp::pparams}.
}
\description{
Convert \code{coef.spatPomp} to \code{pparams} format
}
