% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pfilter_dfs.R
\name{tidy_pfilter_dfs}
\alias{tidy_pfilter_dfs}
\title{Create tidy data frame from saved \code{eval_logLik()} output.}
\usage{
tidy_pfilter_dfs(EL_list)
}
\arguments{
\item{EL_out}{List shaped like the output of \code{eval_logLik()}.}
}
\value{
Tibble of \code{EL_out} data in tidy format.
}
\description{
Create tidy data frame from saved \code{eval_logLik()} output.
}
\examples{
model_list = list(AK_model())
EL_out = eval_logLik(model_list, ncores = 1, np_pf = 3, nreps = 2)
tidy_pfilter_dfs(EL_out)
}
